# tests/unit/data/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutfitRequest": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "OutfitRequest",
        "type": "object"
      },
      "PortraitEditRequest": {
        "properties": {
          "art_style": {
            "default": "children's story book",
            "title": "Art Style",
            "type": "string"
          },
          "mode": {
            "default": "edit",
            "title": "Mode",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "use_current_as_ref": {
            "default": false,
            "title": "Use Current As Ref",
            "type": "boolean"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "PortraitEditRequest",
        "type": "object"
      },
...
      "SceneEditRequest": {
        "properties": {
          "prompt": {
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "SceneEditRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "par-storygen API"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/images/{game_id}/cover/regenerate": {
      "post": {
        "description": "Regenerate the cover art (root node scene image).",
        "operationId": "regenerate_cover_api_images__game_id__cover_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Regenerate Cover Api Images  Game Id  Cover Regenerate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Cover",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}": {
      "get": {
        "description": "Serve the latest character portrait PNG.",
        "operationId": "get_portrait_image_api_images__game_id__portrait__char_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Portrait Image",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/edit": {
      "post": {
        "description": "Edit a character portrait prompt and regenerate.",
        "operationId": "edit_portrait_api_images__game_id__portrait__char_id__edit_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortraitEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Edit Portrait Api Images  Game Id  Portrait  Char Id  Edit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Portrait",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/outfit": {
      "post": {
        "description": "Create a new outfit for a character and generate its portrait.",
        "operationId": "add_outfit_api_images__game_id__portrait__char_id__outfit_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutfitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Outfit Api Images  Game Id  Portrait  Char Id  Outfit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Outfit",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/outfit/revert": {
      "post": {
        "description": "Revert to the base portrait (unset current outfit).",
        "operationId": "revert_outfit_api_images__game_id__portrait__char_id__outfit_revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Revert Outfit Api Images  Game Id  Portrait  Char Id  Outfit Revert Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert Outfit",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/outfit/{outfit_id}": {
      "delete": {
        "description": "Delete an outfit.",
        "operationId": "delete_outfit_api_images__game_id__portrait__char_id__outfit__outfit_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outfit_id",
            "required": true,
            "schema": {
              "title": "Outfit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Outfit Api Images  Game Id  Portrait  Char Id  Outfit  Outfit Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Outfit",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/outfit/{outfit_id}/set": {
      "post": {
        "description": "Set an outfit as the current active outfit.",
        "operationId": "set_outfit_api_images__game_id__portrait__char_id__outfit__outfit_id__set_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "outfit_id",
            "required": true,
            "schema": {
              "title": "Outfit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Set Outfit Api Images  Game Id  Portrait  Char Id  Outfit  Outfit Id  Set Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Outfit",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/portrait/{char_id}/retry": {
      "post": {
        "description": "Retry (regenerate) a character portrait using the stored portrait_prompt.",
        "operationId": "retry_portrait_api_images__game_id__portrait__char_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "title": "Char Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Retry Portrait Api Images  Game Id  Portrait  Char Id  Retry Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Portrait",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/scene/{node_id}": {
      "get": {
        "description": "Serve a scene PNG for a story node.",
        "operationId": "get_scene_image_api_images__game_id__scene__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene Image",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{game_id}/scene/{node_id}/edit": {
      "post": {
        "description": "Edit scene prompt and regenerate.",
        "operationId": "edit_scene_api_images__game_id__scene__node_id__edit_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Edit Scene Api Images  Game Id  Scene  Node Id  Edit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Scene",
        "tags": [
          "images"
        ]
      }
    },
...
  }
}